+++ src/prioritized_experience_replay.jl
@@ -29,10 +29,10 @@ mutable struct PrioritizedReplayBuffer
         s_dim = obs_dimensions(env)
         experience = Vector{DQExperience}(undef, max_size)
         priorities = Vector{Float64}(undef, max_size)
-        _s_batch = zeros(batch_size, s_dim...)
+        _s_batch = zeros(s_dim..., batch_size)
         _a_batch = zeros(Int64, batch_size)
         _r_batch = zeros(batch_size)
-        _sp_batch = zeros(batch_size, s_dim...)
+        _sp_batch = zeros(s_dim..., batch_size)
         _done_batch = zeros(Bool, batch_size)
         _weights_batch = zeros(Float64, batch_size)
         return new(max_size, batch_size, rng, α, β, ϵ, 0, 1, priorities, experience,
@@ -72,10 +72,10 @@ end
 function get_batch(r::PrioritizedReplayBuffer, sample_indices::Vector{Int64})
     @assert length(sample_indices) == size(r._s_batch)[1]
     for (i, idx) in enumerate(sample_indices)
-        r._s_batch[Base.setindex(axes(r._s_batch), i, 1)...] = r._experience[idx].s
+        r._s_batch[Base.setindex(axes(r._s_batch), i, ndims(r._s_batch))...] = r._experience[idx].s
         r._a_batch[i] = r._experience[idx].a
         r._r_batch[i] = r._experience[idx].r
-        r._sp_batch[Base.setindex(axes(r._sp_batch), i, 1)...] = r._experience[idx].sp
+        r._sp_batch[Base.setindex(axes(r._sp_batch), i, ndims(r._sp_batch))...] = r._experience[idx].sp
         r._done_batch[i] = r._experience[idx].done
         r._weights_batch[i] = r._priorities[idx]
     end
