+++ test/test_env.jl
@@ -1,3 +1,4 @@
+using Random
 using POMDPs
 using POMDPModelTools
 using DeepRL
